Create ISR LED Counter with Debounce
Using Assembly, you will write a program that performs various interrupt tasks without any continuous code. When the left button is not pressed, the red LED needs to blink with a period of 0.5 seconds. When the button is pressed, turn the LED off and increment a counter variable (can be a CPU register). If the right button is pressed the green LED must blink the same number of times that the button was pressed, with a period of 0.5 seconds (reset the counter when done, do not allow the counter to increment while blinking). If no buttons are pressed for 5 seconds, stop all operations and go into low power mode 3. While in low power mode 3, blink both LEDs with a period of 0.5s. If either of the two push buttons are pressed while in low power mode, wake back up and resume operations as usual (blinking the green led should also prevent entering low power mode 3). As stated prior, no “main loop” is allowed to run in your code, all operations must be handled using push button and timer interrupt service routines.