Using Assembly Create a program that will increment and decrement the display and countdown and once it's reached zero blink
Using Assembly, you write a program which utilizes the 4x7 segment daughter board to implement a timer. Initially the display should read “00:00”, where the left 2 digits are seconds, and the right 2 digits are milliseconds. When the left button (on the daughter board) is pressed, increment the timer by 1 second. When the middle button (on the daughter board) is pressed, decrease the timer by 1 second. When the right button (on the daughter board) is pressed, start the countdown timer. The timer should countdown in one millisecond at a time, until the timer reaches zero. When the timer reaches zero, blink the seven-segment display with a period of 0.5s. If the left button or middle button is pressed while it is counting down, the timer should pause. The timer should not be allowed to go below zero.

Entire project must be contained within multiple interrupt service routines with no “mainloop” code and must use a look up table for the BCD to segment conversion.